+++ app/Http/Controllers/BancaController.php
@@ -17,15 +17,18 @@ class BancaController extends Controller
     
     public function store(BancaRequest $request)
     {
-        //$this->authorize('admin');
+        $this->authorize('LOGADO');
         $validated = $request->validated();
+        if($validated['codpes'] != ''){
+            $validated['nome'] = Pessoa::dump($validated['codpes'])['nompes'];
+        }
         Banca::create($validated);
         return back();
     }
 
     public function destroy(Banca $banca)
     {
-        //$this->authorize('admin');
+        $this->authorize('LOGADO');
         $banca->delete();
         return back();
     }
